remove first hill on tokyo 2400 for global
unclear if this hill actually exists in global currently. we cheat by only using post-first-anni course data (after the slope change) so the existing course data isn't a perfect match. observations suggest slope==1 skills don't activate on the gate hill, so remove this as a temporary fix because this track is very important. not too happy with this solution.